requirements.txt
python-novaclient
python-neutronclient
jinja2
PyYaml
